% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_package.R
\name{content}
\alias{content}
\title{Example dataset}
\format{A dataframe with 8 variables:
\describe{
  \item{Y}{Observed HAZ score}
  \item{Ytrue}{True HAZ score}
  \item{waz.true}{True WAZ score}
  \item{waz}{Observed WAZ score}
  \item{Male}{Sex. 1 if male, 0 if female.}
  \item{argvals}{time of observations standardized to be in the interval [0,1]}
  \item{subj}{Subject ID}
  \item{include}{Indicator for out of sample prediction used in the vignette}
}}
\description{
Simulated data from the CONTENT dataset. Data contains information on child growth
as measured by WHO defined Z-scores
as well as gender.
}
\references{
Jaganath D, Saito M Giman RH Queirox DM, Rocha GA, Cama V, Cabrera L, Kelleher D, Windle HJ,
Crabtree JE, Jean E, Checkley W. First Detected Helicobacter pylori Infection in Infancy Modifies the
Association Between Diarrheal Disease and Childhood Growth in Peru. Helicobacter (2014); 19:272-297.
}
