% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_methods.R
\name{plot.fcr}
\alias{plot.fcr}
\title{Plotting an fcr model fit}
\usage{
\method{plot}{fcr}(x, plot.covariance = FALSE, ...)
}
\arguments{
\item{x}{object of class fcr.}

\item{plot.covariance}{logical argument, indicates whether to plot features of the covariance
function (correlation function, variance function, and eigenfunctions of the covariance function).
If FALSE, will call \code{\link{plot.gam}} on the fitted gam/bam object. See \code{\link{plot.gam}} for additional details.
Defaults to FALSE.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}
}
\value{
If plot.covariance is FALSE, this function will silently return a list fo the data used to create the plots.
}
\description{
Plot method for \code{\link{fcr}}. Takes a fitted fcr object and plots either the features of the covariance function, or the
smooth terms and qqplots for random effects. See \code{\link{plot.gam}} for further details.
}
\examples{

## see examples in fcr

}
