% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr-package.R
\docType{package}
\name{fcr-package}
\alias{fcr}
\alias{fcr-package}
\title{fcr: Forecasting Cash Rate}
\description{
Forecasting Cash Rate monthly using publicly available data and open-source software, and facilitating reproducible workflows
}
\author{
\strong{Maintainer}: Tomasz Woźniak \email{wozniak.tom@pm.me} (\href{https://orcid.org/0000-0003-2212-2378}{ORCID})

}
\keyword{internal}
